// src/interfaces/delivered_items.interface.ts
export default interface DeliveredItemsInterface {
  delivered_item_id: string;
  unique_receiver_id: string;
  entity_id: string;
  staff_id: string;
  ik_number: string;
  item_id: string;
  delivery_date: string;
  app_version: string;
  imei: string;
  hub_id: string;
  item_info: string;
  item_amount: number;
  reason: string;
  picture_name: string;
  delivery_id: string;
  sync_flag: number;
}